requirements.txt
# Core ML
torch==2.2.0
torchvision==0.17.0
timm==0.9.12

# Data & Processing
numpy==1.26.4
pandas==2.2.0
Pillow==10.2.0
opencv-python==4.9.0.80
scikit-learn==1.4.0
albumentations==1.3.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0

# Experiment Tracking
mlflow==2.10.0

# Web App
streamlit==1.31.0
streamlit-extras==0.4.0

# Utilities
tqdm==4.66.2
pyyaml==6.0.1
python-dotenv==1.0.1
kaggle==1.6.3
grad-cam==1.5.4

# Testing
pytest==8.0.2
